Create Marathon
Chefland is holding a virtual marathon for the categories 10 km, 21 km and 42 km having prizes ,A,B,C (A<B<C) respectively to promote physical fitness. In order to claim the prize in a particular category the person must cover the total distance for that category within D days. Also a single person cannot apply in multiple categories.
Given the maximum distance d km that Chef can cover in a single day, find the maximum prize that Chef can win at the end of D days. If Chef can't win any prize, print 0.
eg:
10 3 1 2 3
op: 2
The maximum distance covered by Chef is 10⋅3=30 km which is more than the distance of the second category but less than that of the third category. Hence Chef can claim the maximum prize of 
2 units.