Create is a number multiplied by 2 divisible by 10
Given a number n. if the number is already divisible by 10 print 0.
Else if a number is not divisible by 10 multiply it by 2.The goal of the game is to make n divisible by 10.
Find the smallest number of turns necessary to win the game (it may be possible to win in zero turns) or determine that it is impossible.
the minimum required number of turns or −1 if there is no way to win the game.