Create non negative product
She wants the product of all the elements of the array to be a non-negative integer. That is, it can be either 0 or positive. But it shouldn't be negative. To do this, we should  remove some elements of the array. Determine the minimum number of elements that will have to  be removed to make the product of the array's elements non-negative.
eg: 2 -1 9 100
op: 1
eg: 2 -1 0 100
op: 0