Create deleting chars to form a palindrome
You are given a string S of length N.
Your task is to delete a subsequence of maximum length from the string, such that, after concatenating the remaining parts of the string, it becomes a palindrome of length greater than 1.
If this is possible, print the maximum length of the subsequence that can be deleted. Otherwise, print −1.
eg: babkhj
op: 4
Possible ways to delete a subsequence are:
Delete subsequence khj to get palindrome bab.
Delete subsequence akhj to get palindrome bb.
The subsequence having maximum length that can be deleted is akhj, having length 4.