Create repeated string
There is a string, , of lowercase English letters that is repeated infinitely many times. Given an integer, , find and print the number of letter a's in the first  letters of the infinite string.

Example
s='abcac'
n=10
The substring we consider is ,abcacabcac the first 10  characters of the infinite string. There are 4 occurrences of a in the substring.