Create Maximum distance from city to its nearest space station.
Flatland is a country with a number of cities, some of which have space stations. Cities are numbered consecutively and each has a road of 1km  length connecting it to the next city. It is not a circular route, so the first city doesn't connect with the last city. Determine the maximum distance from any city to its nearest space station.

Sample Input 0

STDIN   Function
-----   --------
5 2     n = 5, c[] size m = 2
0 4     c = [0, 4]
Sample Output 0

2